Fixed integer division

+++ ContactTracing15.Services/Repository Interfaces/CaseService.cs
@@ -94,7 +94,7 @@ namespace ContactTracing15.Services
             int cases = _caseRepository.GetCasesByDate(DateTime.Now, DateTime.Now.AddDays(-28)).ToList().Count();
             if (cases == 0) { return 0; }
             int traced = _caseRepository.GetCasesByDate(DateTime.Now, DateTime.Now.AddDays(-28)).Where(x => x.Traced).ToList().Count();
-            return traced / cases * 100;
+            return  (double) traced / cases * 100;
         }
